Fix warnings + Better doc
